dataset: address XM3600 I2T review feedback
