+++ mteb/tasks/retrieval/multilingual/__init__.py
@@ -142,8 +142,7 @@ from .wit_t2i_retrieval import WITT2IRetrieval
 from .x_flickr30k_co_t2i_retrieval import XFlickr30kCoT2IRetrieval
 from .x_market_retrieval import XMarket
 from .x_qu_ad_retrieval import XQuADRetrieval
-from .xm3600_i2t_retrieval import XM3600I2TRetrieval
-from .xm3600_t2i_retrieval import XM3600T2IRetrieval
+from .xm3600_t2i_retrieval import XM3600I2TRetrieval, XM3600T2IRetrieval
 from .xpqa_retrieval import XPQARetrieval
 
 __all__ = [
